Removed reference to CI/CD pipeline test and reorganised folders to reduce the number of small mod.rs files
