--- src/radiometry/mod.rs
@@ -1,2 +1,0 @@
-pub mod spectra;
-pub mod constants;
